// client/src/generated/api-spec.d.ts
export interface components {
  schemas: {
    /**
     * Format: uuid
     * @example 00000000-0000-0000-0000-000000000000
     */
    UUID: string;
    /**
     * @example {
     *   "puzzle": "00000000-0000-0000-0000-000000000000",
     *   "submittedAt": null,
     *   "title": "Lauren Ipsum",
     *   "widgets": {}
     * }
     */
    Blueprint: {
      puzzle: string;
      submittedAt?: components["schemas"]["UTCTime"];
      title: string;
      widgets: components["schemas"]["Object"];
    };
    /**
     * Format: yyyy-mm-ddThh:MM:ssZ
     * @example 2016-07-22T00:00:00Z
     */
    UTCTime: string;
    /** @description Arbitrary JSON object. */
    Object: {
      [key: string]: unknown;
    };
    Folio: {
      blueprint: components["schemas"]["Blueprint"];
      puzzle: components["schemas"]["Puzzle"];
      snapshot: components["schemas"]["Object"];
    };
    /**
     * @example {
     *   "inputs": [
     *     {
     *       "balls": [
     *         {
     *           "rate": 1,
     *           "type": 1
     *         }
     *       ],
     *       "x": 0.5,
     *       "y": 0
     *     }
     *   ],
     *   "outputs": [
     *     {
     *       "balls": [
     *         {
     *           "rate": 1,
     *           "type": 1
     *         }
     *       ],
     *       "x": 0.5,
     *       "y": 1
     *     }
     *   ],
     *   "reqTiles": [
     *     "UpLeft"
     *   ],
     *   "spec": {}
     * }
     */
    Puzzle: {
      inputs: components["schemas"]["Gateway"][];
      outputs: components["schemas"]["Gateway"][];
      reqTiles: components["schemas"]["RelativeCell"][];
      spec: components["schemas"]["Object"];
    };
    /** @enum {string} */
    RelativeCell: "UpLeft" | "Up" | "UpRight" | "Left" | "Right" | "DownLeft" | "Down" | "DownRight";
    /**
     * @example {
     *   "blueprint": "00000000-0000-0000-0000-000000000000",
     *   "puzzle": "00000000-0000-0000-0000-000000000000",
     *   "to_mod": 20
     * }
     */
    Gateway: {
      balls: components["schemas"]["GatewayBall"][];
      /** Format: double */
      x: number;
      /** Format: double */
      y: number;
    };
    GatewayBall: {
      /** Format: double */
      rate: number;
      type: number;
    };
    /**
     * @example {
     *   "grid": [
     *     [
     *       "00000000-0000-0000-0000-000000000000",
     *       "00000000-0000-0000-0000-000000000000"
     *     ],
     *     [
     *       "00000000-0000-0000-0000-000000000000",
     *       null
     *     ]
     *   ],
     *   "ms_per_ball": 0.001,
     *   "prio_puzzles": [],
     *   "tile_size": {
     *     "x": 700,
     *     "y": 700
     *   },
     *   "version": 0
     * }
     */
    "VersionedMachine (Maybe BlueprintID)": {
      grid: components["schemas"]["UUID"][][];
      /** Format: double */
      ms_per_ball: number;
      prio_puzzle?: components["schemas"]["UUID"][];
      tile_size: components["schemas"]["TileSize"];
      version: number;
    };
    TileSize: {
      x: number;
      y: number;
    };
    /**
     * @example {
     *   "grid": [
     *     [
     *       {
     *         "blueprint": "00000000-0000-0000-0000-000000000000",
     *         "puzzle": "00000000-0000-0000-0000-000000000000"
     *       },
     *       {
     *         "puzzle": "00000000-0000-0000-0000-000000000000",
     *         "to_mod": 20
     *       }
     *     ],
     *     [
     *       {
     *         "puzzle": "00000000-0000-0000-0000-000000000000",
     *         "to_mod": 11
     *       },
     *       {
     *         "puzzle": "00000000-0000-0000-0000-000000000000"
     *       }
     *     ]
     *   ],
     *   "ms_per_ball": 0.001,
     *   "prio_puzzles": [],
     *   "tile_size": {
     *     "x": 700,
     *     "y": 700
     *   },
     *   "version": 0
     * }
     */
    "VersionedMachine ModData": {
      grid: components["schemas"]["ModData"][][];
      /** Format: double */
      ms_per_ball: number;
      prio_puzzle?: components["schemas"]["UUID"][];
      tile_size: components["schemas"]["TileSize"];
      version: number;
    };
    InspectionReport: {
      blueprint: components["schemas"]["UUID"];
      snapshot: components["schemas"]["Object"];
    };
    /**
     * @example {
     *   "blueprint": "00000000-0000-0000-0000-000000000000",
     *   "puzzle": "00000000-0000-0000-0000-000000000000",
     *   "to_mod": 20
     * }
     */
    ModData: {
      blueprint?: string;
      puzzle: string;
      to_mod?: number;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}